refactor: switch BookmarkLanding to SendToWallet
