+++ packages/widget/src/components/BookmarkLanding/index.ts
@@ -1,1 +1,1 @@
-export * from './BookmarkLanding'
+export * from './SendToWallet';
